Modify op_benchmark directory structure to add bench_tests/ and bench_model.py
* Why? The idea is to have a master script that benchmarks the full set of associated kernels when given a model name. It's a little cleaner to place all kernel benchmarking scripts in /kernels and have the bench_model script
call them.
* How? See `bench_model.py`. Pytests are in bench_tests/
