+++ op_tests/op_benchmarks/triton/kernels/bench_batched_gemm_afp4wfp4.py
@@ -2,14 +2,17 @@ import argparse
 import sys
 import torch
 import triton
-from aiter.ops.triton.gemm_a8w8_blockscale import gemm_a8w8_blockscale
-from op_tests.triton_tests.test_gemm_a8w8_blockscale import (
-    generate_gemm_a8w8_blockscale_inputs,
+from op_tests.triton_tests.test_batched_gemm_afp4wfp4 import (
+    generate_batched_gemm_afp4wfp4_inputs,
 )
-from utils.benchmark_utils import get_model_configs, get_available_models
+from op_tests.op_benchmarks.triton.utils.benchmark_utils import (
+    get_model_configs,
+    get_available_models,
+)
 
-
-block_shape = (128, 128)
+from aiter.ops.triton.batched_gemm_afp4wfp4 import (
+    batched_gemm_afp4wfp4 as batched_gemm_afp4wfp4,
+)
 
 
 def model_benchmark_shapes(args):
@@ -22,26 +25,26 @@ def model_benchmark_shapes(args):
             N = config["intermediate_size"]
             K = config["hidden_size"]
 
-            shapes.append((M, N, K))
+            shapes.append((16, M, N, K))
 
     return shapes
 
 
 def get_x_vals():
     x_vals = [
-        (1, 1280, 8192),
-        (32, 1280, 8192),
-        (64, 1280, 8192),
-        (128, 1280, 8192),
-        (192, 1280, 8192),
-        (256, 1280, 8192),
-        (320, 1280, 8192),
-        (512, 1280, 8192),
-        (1024, 1280, 8192),
-        (2048, 1280, 8192),
-        (4096, 1280, 8192),
-        (8192, 1280, 8192),
-        (16384, 1280, 8192),
+        (16, 1, 1280, 8192),
+        (16, 32, 1280, 8192),
+        (16, 64, 1280, 8192),
+        (16, 128, 1280, 8192),
+        (16, 192, 1280, 8192),
+        (16, 256, 1280, 8192),
+        (16, 320, 1280, 8192),
+        (16, 512, 1280, 8192),
+        (16, 1024, 1280, 8192),
+        (16, 2048, 1280, 8192),
+        (16, 4096, 1280, 8192),
+        (16, 8192, 1280, 8192),
+        (16, 16384, 1280, 8192),
     ]
     return x_vals
 
@@ -51,7 +54,7 @@ def run_benchmark(args):
         args.shape and args.M
     ), "User can specify --shape or --model MODEL -M VAL exclusively"
 
-    x_names = ["M", "N", "K"]
+    x_names = ["batch", "M", "N", "K"]
     if args.model:
         x_vals_list = model_benchmark_shapes(args)
     elif args.shape:
@@ -78,29 +81,30 @@ def run_benchmark(args):
         line_names=line_names,
         styles=[("green", "-")],
         ylabel=ylabel,
-        plot_name="GEMM A8W8 Benchmark",
+        plot_name="GEMM MXFP4 x MXFP4 Benchmark",
         args={"metric": args.metric},
     )
 
     @triton.testing.perf_report([benchmark])
-    def bench_gemm_a8w8_blockscale(M, N, K, metric, provider):
-        block_shape_n, block_shape_k = block_shape
-
+    def bench_gemm_afp4wfp4_blockscale(batch, M, N, K, metric, provider):
         c_dtype = torch.bfloat16
-        x, weight, x_scale, w_scale, y = generate_gemm_a8w8_blockscale_inputs(
-            M, N, K, block_shape_n, block_shape_k, output=True
-        )
+        x, w, x_scale, w_scale = generate_batched_gemm_afp4wfp4_inputs(batch, M, N, K)
         # flops
         flops = 2.0 * M * N * K
         # memory transfer
-        mem_read = (M * K) * x.element_size() + (N * K) * weight.element_size()
+        mem_read = x.numel() * x.element_size() + w.numel() * w.element_size()
+        mem_read += (
+            x_scale.numel() * x_scale.element_size()
+            + w_scale.numel() * w_scale.element_size()
+        )
         mem_write = (M * N) * 2  # TODO: Fix for c_dtype != bf16
         mem = mem_read + mem_write
+        out = torch.empty(
+            x.shape[0], x.shape[1], w.shape[2], device=x.device, dtype=c_dtype
+        )
 
         ms = triton.testing.do_bench(
-            lambda: gemm_a8w8_blockscale(
-                x, weight, x_scale, w_scale, c_dtype, y
-            ),  # noqa: E731
+            lambda: batched_gemm_afp4wfp4(x, w, x_scale, w_scale, c_dtype, out),
             warmup=25,
             rep=100,
         )
@@ -117,12 +121,12 @@ def run_benchmark(args):
         else:
             raise ValueError("Unknown metric: " + metric)
 
-    bench_gemm_a8w8_blockscale.run(save_path=".", print_data=True)
+    bench_gemm_afp4wfp4_blockscale.run(save_path=".", print_data=True)
 
 
 def parse_args():
     parser = argparse.ArgumentParser(
-        prog="Benchmark A8W8 GEMM",
+        prog="Benchmark MXFP4 x MXFP4 GEMM",
         allow_abbrev=False,
         formatter_class=argparse.ArgumentDefaultsHelpFormatter,
     )
@@ -148,8 +152,8 @@ def parse_args():
     parser.add_argument(
         "--shape",
         type=int,
-        nargs=3,
-        metavar=("M", "N", "K"),
+        nargs=4,
+        metavar=("batch", "M", "N", "K"),
         help="user-defined shape to benchmark",
     )
     parser.add_argument(
