+++ op_tests/op_benchmarks/triton/kernels/bench_batched_gemm_afp4wfp4_pre_quant.py
@@ -2,9 +2,17 @@ import argparse
 import sys
 import torch
 import triton
-from aiter.ops.triton.rmsnorm import rms_norm
-from op_tests.triton_tests.test_rmsnorm import generate_rmsnorm_inputs
-from utils.benchmark_utils import get_model_configs, get_available_models
+from op_tests.triton_tests.test_batched_gemm_afp4wfp4_pre_quant import (
+    generate_batched_gemm_afp4wfp4_pre_quant_inputs,
+)
+from op_tests.op_benchmarks.triton.utils.benchmark_utils import (
+    get_model_configs,
+    get_available_models,
+)
+
+from aiter.ops.triton.batched_gemm_afp4wfp4_pre_quant import (
+    batched_gemm_afp4wfp4_pre_quant as batched_gemm_afp4wfp4_pre_quant,
+)
 
 
 def model_benchmark_shapes(args):
@@ -14,29 +22,29 @@ def model_benchmark_shapes(args):
     shapes = []
     for M in M_list:
         for _, config in configs.items():
-            num_heads = config["num_attention_heads"]
-            N = config["hidden_size"]
+            N = config["intermediate_size"]
+            K = config["hidden_size"]
 
-            shapes.append((M, N))
+            shapes.append((16, M, N, K))
 
     return shapes
 
 
 def get_x_vals():
     x_vals = [
-        (1, 1280),
-        (32, 1280),
-        (64, 1280),
-        (128, 1280),
-        (192, 1280),
-        (256, 1280),
-        (320, 1280),
-        (512, 1280),
-        (1024, 1280),
-        (2048, 1280),
-        (4096, 1280),
-        (8192, 1280),
-        (16384, 1280),
+        (16, 1, 1280, 8192),
+        (16, 32, 1280, 8192),
+        (16, 64, 1280, 8192),
+        (16, 128, 1280, 8192),
+        (16, 192, 1280, 8192),
+        (16, 256, 1280, 8192),
+        (16, 320, 1280, 8192),
+        (16, 512, 1280, 8192),
+        (16, 1024, 1280, 8192),
+        (16, 2048, 1280, 8192),
+        (16, 4096, 1280, 8192),
+        (16, 8192, 1280, 8192),
+        (16, 16384, 1280, 8192),
     ]
     return x_vals
 
@@ -46,7 +54,7 @@ def run_benchmark(args):
         args.shape and args.M
     ), "User can specify --shape or --model MODEL -M VAL exclusively"
 
-    x_names = ["M", "N"]
+    x_names = ["batch", "M", "N", "K"]
     if args.model:
         x_vals_list = model_benchmark_shapes(args)
     elif args.shape:
@@ -56,6 +64,8 @@ def run_benchmark(args):
 
     if args.metric == "time":
         ylabel = "Time (ms)"
+    elif args.metric == "throughput":
+        ylabel = "Throughput (TFLOPS)"
     elif args.metric == "bandwidth":
         ylabel = "Bandwidth (GB/s)"
     else:
@@ -71,37 +81,56 @@ def run_benchmark(args):
         line_names=line_names,
         styles=[("green", "-")],
         ylabel=ylabel,
-        plot_name="RMSNorm Fwd Benchmark",
+        plot_name="GEMM MXFP4 x MXFP4 Benchmark",
         args={"metric": args.metric},
     )
 
     @triton.testing.perf_report([benchmark])
-    def bench_rmsnorm(M, N, metric, provider):
+    def bench_gemm_afp4wfp4_blockscale(batch, M, N, K, metric, provider):
         c_dtype = torch.bfloat16
-        x, w = generate_rmsnorm_inputs(M, N, c_dtype)
-
+        x, w, x_scale, w_scale = generate_batched_gemm_afp4wfp4_pre_quant_inputs(
+            batch, M, N, K
+        )
+        # flops
+        flops = 2.0 * M * N * K
         # memory transfer
-        mem_read = (M * 1) * N * x.element_size()  # x is (M,N) and g/weight is (N)
-        mem_write = M * N * x.element_size()  # output
+        mem_read = x.numel() * x.element_size() + w.numel() * w.element_size()
+        mem_read += (
+            x_scale.numel() * x_scale.element_size()
+            + w_scale.numel() * w_scale.element_size()
+        )
+        mem_write = (M * N) * 2  # TODO: Fix for c_dtype != bf16
         mem = mem_read + mem_write
+        out = torch.empty(
+            x.shape[0], x.shape[1], w.shape[2], device=x.device, dtype=c_dtype
+        )
 
-        ms = triton.testing.do_bench(lambda: rms_norm(x, w), warmup=25, rep=100)
+        ms = triton.testing.do_bench(
+            lambda: batched_gemm_afp4wfp4_pre_quant(
+                x, w, x_scale, w_scale, c_dtype, out
+            ),
+            warmup=25,
+            rep=100,
+        )
 
         # Return exactly one scalar depending on which metric is active
         if metric == "time":
             return ms
+        elif metric == "throughput":
+            tflops = flops / ms * 1e-9
+            return tflops
         elif metric == "bandwidth":
             bandwidth = mem / (ms * 1e-3) * 1e-9  # GB/s
             return bandwidth
         else:
             raise ValueError("Unknown metric: " + metric)
 
-    bench_rmsnorm.run(save_path=".", print_data=True)
+    bench_gemm_afp4wfp4_blockscale.run(save_path=".", print_data=True)
 
 
 def parse_args():
     parser = argparse.ArgumentParser(
-        prog="Benchmark RMSNorm",
+        prog="Benchmark MXFP4 x MXFP4 GEMM",
         allow_abbrev=False,
         formatter_class=argparse.ArgumentDefaultsHelpFormatter,
     )
@@ -127,15 +156,15 @@ def parse_args():
     parser.add_argument(
         "--shape",
         type=int,
-        nargs=2,
-        metavar=("M", "N"),
+        nargs=4,
+        metavar=("batch", "M", "N", "K"),
         help="user-defined shape to benchmark",
     )
     parser.add_argument(
         "--metric",
         type=str,
-        choices=["time", "bandwidth"],
-        default="bandwidth",
+        choices=["time", "throughput", "bandwidth"],
+        default="throughput",
         help="metric to plot",
     )
     args = parser.parse_args()
