+++ op_tests/op_benchmarks/triton/kernels/bench_gemm_a16w16.py
@@ -4,7 +4,10 @@ import torch
 import triton
 from aiter.ops.triton.gemm_a16w16 import gemm_a16w16
 from op_tests.triton_tests.test_gemm_a16w16 import generate_gemm_a16w16_inputs
-from utils.benchmark_utils import get_model_configs, get_available_models
+from op_tests.op_benchmarks.triton.utils.benchmark_utils import (
+    get_model_configs,
+    get_available_models,
+)
 
 
 def model_benchmark_shapes(args):
