+++ op_tests/op_benchmarks/triton/kernels/bench_moe.py
@@ -2,11 +2,10 @@ import sys
 import argparse
 import torch
 import triton
-import triton.language as tl
 from aiter.ops.triton.utils.types import torch_to_triton_dtype, str_to_torch_dtype
 from aiter.ops.triton.moe_op import fused_moe as triton_moe
 from op_tests.triton_tests.test_moe import input_helper, input_helper_int4_w4a16
-from utils.benchmark_utils import (
+from op_tests.op_benchmarks.triton.utils.benchmark_utils import (
     get_model_configs,
     get_available_models,
 )
