+++ op_tests/op_benchmarks/triton/kernels/bench_moe_mx.py
@@ -9,7 +9,10 @@ from op_tests.triton_tests.test_moe_mx import (
     alloc_rand,
     torch_dynamic_mxfp4_quant,
 )
-from utils.benchmark_utils import get_available_models, get_model_configs
+from op_tests.op_benchmarks.triton.utils.benchmark_utils import (
+    get_available_models,
+    get_model_configs,
+)
 
 
 def model_benchmark_configs(args):
