+++ op_tests/op_benchmarks/triton/kernels/bench_pa_decode.py
@@ -3,9 +3,8 @@ import random
 import torch
 import argparse
 import triton
-import triton.language as tl
 from aiter.ops.triton.pa_decode import paged_attention_decode
-from utils.benchmark_utils import (
+from op_tests.op_benchmarks.triton.utils.benchmark_utils import (
     get_model_configs,
     get_available_models,
     get_dtype_bytes,
