+++ op_tests/op_benchmarks/triton/kernels/bench_pa_prefill.py
@@ -1,6 +1,5 @@
 import triton
-import triton.language as tl
-from utils.benchmark_utils import (
+from op_tests.op_benchmarks.triton.utils.benchmark_utils import (
     get_model_configs,
     get_available_models,
     get_dtype_bytes,
